backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4
statsmodels==0.14.0
openai==1.3.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
boto3==1.34.0
openpyxl==3.1.2
xlrd==2.0.1
pyreadstat==1.2.4
